# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: The Storyteller's Stage – API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Auth status resolved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the OIDC authorization endpoint
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to returnTo on success
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
            default: /
      responses:
        "302":
          description: Redirect to OIDC logout

  # ─── Storage ──────────────────────────────────────────────────────────────
...
  /admin/plays:
    get:
      tags: [Admin]
      operationId: listAdminPlays
      summary: List all plays with translation status
      responses:
        "200":
          description: All plays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminPlaySummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      tags: [Admin]
      operationId: createAdminPlay
      summary: Create a new play
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayBaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPlayDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/plays/{id}:
    get:
      tags: [Admin]
      operationId: getAdminPlay
      summary: Get a play with all its translations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Play with all translations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPlayDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      tags: [Admin]
      operationId: updateAdminPlay
      summary: Update play base metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayBaseInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPlayDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [Admin]
      operationId: deleteAdminPlay
      summary: Delete a play and all its translations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/plays/{id}/translations/{locale}:
    put:
      tags: [Admin]
      operationId: upsertPlayTranslation
      summary: Create or update a translation for a play
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: locale
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayTranslationInput"
      responses:
        "200":
          description: Translation saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPlayDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/plays/{id}/translations/{locale}/publish:
    post:
      tags: [Admin]
      operationId: publishPlayTranslation
      summary: Publish a play translation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: locale
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPlayDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Translation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/plays/{id}/translations/{locale}/unpublish:
    post:
      tags: [Admin]
      operationId: unpublishPlayTranslation
      summary: Unpublish a play translation (back to draft)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: locale
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unpublished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPlayDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Translation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ─── Admin – About ────────────────────────────────────────────────────────
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: "Opaque session token — Bearer <sid>"
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, "null"]
        firstName:
          type: [string, "null"]
        lastName:
          type: [string, "null"]
        profileImageUrl:
          type: [string, "null"]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"

    # ── Storage ───────────────────────────────────────────────────────────────
...
    TranslationStatus:
      type: object
      required: [locale, status]
      properties:
        locale:
          type: string
        status:
          type: string
          description: "draft | published | archived | missing"
        publishedAt:
          type: [string, "null"]
          format: date-time

    # ── Fairy Tales – Admin ───────────────────────────────────────────────────
...
    AdminPlaySummary:
      type: object
      required: [id, slug, createdAt, updatedAt, translations]
      properties:
        id:
          type: number
        slug:
          type: string
        genre:
          type: [string, "null"]
        targetAudience:
          type: [string, "null"]
        estimatedDuration:
          type: [number, "null"]
        castSize:
          type: [string, "null"]
        scriptAvailability:
          type: string
        stagingComplexity:
          type: [string, "null"]
        productionHistory:
          type: [string, "null"]
        awards:
          type: [string, "null"]
        coverImagePath:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        translations:
          type: array
          items:
            $ref: "#/components/schemas/TranslationStatus"
    AdminPlayTranslation:
      type: object
      required: [id, playId, locale, status, title, createdAt, updatedAt]
      properties:
        id:
          type: number
        playId:
          type: number
        locale:
          type: string
        status:
          type: string
        title:
          type: string
        logline:
          type: [string, "null"]
        synopsis:
          type: [string, "null"]
        excerpt:
          type: [string, "null"]
        stagingNotes:
          type: [string, "null"]
        productionInfo:
          type: [string, "null"]
        seoTitle:
          type: [string, "null"]
        seoDescription:
          type: [string, "null"]
        publishedAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminPlayDetail:
      type: object
      required: [id, slug, scriptAvailability, createdAt, updatedAt, translations]
      properties:
        id:
          type: number
        slug:
          type: string
        genre:
          type: [string, "null"]
        targetAudience:
          type: [string, "null"]
        estimatedDuration:
          type: [number, "null"]
        castSize:
          type: [string, "null"]
        scriptAvailability:
          type: string
        stagingComplexity:
          type: [string, "null"]
        productionHistory:
          type: [string, "null"]
        awards:
          type: [string, "null"]
        coverImagePath:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        translations:
          type: array
          items:
            $ref: "#/components/schemas/AdminPlayTranslation"
    PlayBaseInput:
      type: object
      required: [slug, scriptAvailability]
      properties:
        slug:
          type: string
          minLength: 1
        genre:
          type: [string, "null"]
        targetAudience:
          type: [string, "null"]
        estimatedDuration:
          type: [number, "null"]
        castSize:
          type: [string, "null"]
        scriptAvailability:
          type: string
        stagingComplexity:
          type: [string, "null"]
        productionHistory:
          type: [string, "null"]
        awards:
          type: [string, "null"]
        coverImagePath:
          type: [string, "null"]
    PlayTranslationInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        logline:
          type: [string, "null"]
        synopsis:
          type: [string, "null"]
        excerpt:
          type: [string, "null"]
        stagingNotes:
          type: [string, "null"]
        productionInfo:
          type: [string, "null"]
        seoTitle:
          type: [string, "null"]
        seoDescription:
          type: [string, "null"]

    # ── Plays – Public ────────────────────────────────────────────────────────
...